pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "softedibo"
version = "0.1.0"
description = "Soft Robotics Education for Inclusive Bodily Interactions - Robot Hospital platform"
requires-python = ">=3.13"
dependencies = [
    "pyserial>=3.5",
    "PyQt6>=6.5",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
thymio = ["tdm-client>=0.10"]
dev = ["pytest>=7.0", "pytest-qt>=4.2"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
